move SkiddleEvent related classes into skiddle model
